requirements.txt
typing
fastapi[standard]
jinja2
psycopg[binary]
python-dotenv
git+https://github.com/openai/CLIP.git
ftfy
regex
tqdm
Pillow
torch
pandas
boto3